Add tests to show support / non-support of cross-file references.
