+++ tests/_data/config-1.yml
@@ -1,0 +1,3 @@
+c: foo
+m:
+  - 1
