+++ tests/_data/config-2.yml
@@ -1,0 +1,3 @@
+b: ${c}bar
+m:
+  - 2
