+++ tests/_data/config-3.yml
@@ -1,0 +1,3 @@
+a: ${b}baz
+m:
+  - 3
