+++ tests/unit/ConfigurationTest.php
@@ -4,6 +4,8 @@ namespace unit;
 
 use Robo\Robo;
 use Robo\Task\BaseTask;
+use Robo\Config\ConfigProcessor;
+use Robo\Config\YamlConfigLoader;
 
 class ConfigurationTest extends \Codeception\TestCase\Test
 {
@@ -21,6 +23,39 @@ class ConfigurationTest extends \Codeception\TestCase\Test
         verify($taskB->run())->equals('value-b');
     }
 
+    public function testConfigurationWithCrossFileReferences()
+    {
+        $processor = new ConfigProcessor();
+        $loader = new YamlConfigLoader();
+        $processor->extend($loader->load(dirname(__DIR__) . '/_data/config-1.yml'));
+        $processor->extend($loader->load(dirname(__DIR__) . '/_data/config-2.yml'));
+        $processor->extend($loader->load(dirname(__DIR__) . '/_data/config-3.yml'));
+
+        \Robo\Robo::config()->import($processor->export());
+
+        verify(implode(',', \Robo\Robo::config()->get('m')))->equals('3');
+        verify(\Robo\Robo::config()->get('a'))->equals('foobarbaz');
+    }
+
+    public function testConfigurationWithReverseOrderCrossFileReferences()
+    {
+        $processor = new ConfigProcessor();
+        $loader = new YamlConfigLoader();
+        $processor->extend($loader->load(dirname(__DIR__) . '/_data/config-3.yml'));
+        $processor->extend($loader->load(dirname(__DIR__) . '/_data/config-2.yml'));
+        $processor->extend($loader->load(dirname(__DIR__) . '/_data/config-1.yml'));
+
+        \Robo\Robo::config()->import($processor->export());
+
+        verify(implode(',', \Robo\Robo::config()->get('m')))->equals('1');
+
+        if (strpos(\Robo\Robo::config()->get('a'), '$') !== false) {
+            throw new \PHPUnit_Framework_SkippedTestError(
+                'Evaluation of cross-file references in reverse order not supported.'
+            );
+        }
+        verify(\Robo\Robo::config()->get('a'))->equals('foobarbaz');
+    }
 }
 
 class ConfigurationTestTaskA extends BaseTask
