+++ tests/unit/RunnerTest.php
@@ -21,52 +21,6 @@ class RunnerTest extends \Codeception\TestCase\Test
         $this->runner = new \Robo\Runner('\Robo\RoboFileFixture');
     }
 
-    public function testHandleError()
-    {
-        $tmpLevel = error_reporting();
-
-        $this->assertFalse($this->runner->handleError());
-        error_reporting(0);
-        $this->assertTrue($this->runner->handleError());
-
-        error_reporting($tmpLevel);
-    }
-
-    public function testErrorIsHandled()
-    {
-        $tmpLevel = error_reporting();
-
-        // Set error_get_last to a known state.  Note that it can never be
-        // reset; see http://php.net/manual/en/function.error-get-last.php
-        @trigger_error('control');
-        $error_description = error_get_last();
-        $this->assertEquals('control', $error_description['message']);
-        @trigger_error('');
-        $error_description = error_get_last();
-        $this->assertEquals('', $error_description['message']);
-
-        // Set error_reporting to a non-zero value.  In this instance,
-        // 'trigger_error' would abort our test script, so we use
-        // @trigger_error so that execution will continue.  With our
-        // error handler in place, the value of error_get_last() does
-        // not change.
-        error_reporting(E_USER_ERROR);
-        set_error_handler(array($this->runner, 'handleError'));
-        @trigger_error('test error', E_USER_ERROR);
-        $error_description = error_get_last();
-        $this->assertEquals('', $error_description['message']);
-
-        // Set error_reporting to zero.  Now, even 'trigger_error'
-        // does not abort execution.  The value of error_get_last()
-        // still does not change.
-        error_reporting(0);
-        trigger_error('test error 2', E_USER_ERROR);
-        $error_description = error_get_last();
-        $this->assertEquals('', $error_description['message']);
-
-        error_reporting($tmpLevel);
-    }
-
     public function testThrowsExceptionWhenNoContainerAvailable()
     {
         \PHPUnit_Framework_TestCase::setExpectedExceptionRegExp(
@@ -308,118 +262,5 @@ EOT;
         $this->guy->seeInOutput(' [notice] This is a notice log message.');
         $this->guy->seeInOutput(' [debug] This is a debug log message.');
         $this->assertEquals(0, $result);
-    }
-
-    public function testNoOptionsNoConfiguration()
-    {
-        // Run without any config and without any options
-        $argv = ['placeholder', 'test:simple-list'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '1'");
-        $this->guy->seeInOutput("b: '2'");
-    }
-
-    public function testOptionsButNoConfiguration()
-    {
-        // Set one option, but provide no config
-        $argv = ['placeholder', 'test:simple-list', '--b=3'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '1'");
-        $this->guy->seeInOutput("b: '3'");
-    }
-
-    public function testWithConfigurationButNoOptions()
-    {
-        \Robo\Robo::config()->set('command.test.simple-list.options.a', '4');
-        \Robo\Robo::config()->set('command.test.simple-list.options.b', '5');
-
-        $argv = ['placeholder', 'test:simple-list'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '4'");
-        $this->guy->seeInOutput("b: '5'");
-    }
-
-    public function testWithConfigurationAndOptionOverride()
-    {
-        \Robo\Robo::config()->set('command.test.simple-list.options.a', '4');
-        \Robo\Robo::config()->set('command.test.simple-list.options.b', '5');
-
-        $argv = ['placeholder', 'test:simple-list', '--b=6'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '4'");
-        $this->guy->seeInOutput("b: '6'");
-    }
-
-    public function testSettingConfigurationFromCommandOptions()
-    {
-        $argv = ['placeholder', 'test:simple-list', '-D', 'config.key=value'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '1'");
-        $this->guy->seeInOutput("b: '2'");
-
-        $actual = \Robo\Robo::config()->get('config.key');
-        $this->assertEquals('value', $actual);
-    }
-
-    public function testWithConfigLoader()
-    {
-        $loader = new YamlConfigLoader();
-        $loader->load(dirname(__DIR__) . '/_data/robo.yml');
-
-        \Robo\Robo::config()->import($loader->export());
-
-        $argv = ['placeholder', 'test:simple-list'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '12'");
-        $this->guy->seeInOutput("b: '13'");
-    }
-
-    public function testWithConfigLoaderAndCliOverride()
-    {
-        $loader = new YamlConfigLoader();
-        $loader->load(dirname(__DIR__) . '/_data/robo.yml');
-
-        \Robo\Robo::config()->import($loader->export());
-
-        $argv = ['placeholder', 'test:simple-list', '--b=3'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '12'");
-        $this->guy->seeInOutput("b: '3'");
-    }
-
-    public function testWithConfigProcessor()
-    {
-        $processor = new ConfigProcessor();
-        $loader = new YamlConfigLoader();
-        $processor->extend($loader->load(dirname(__DIR__) . '/_data/robo.yml'));
-        $processor->extend($loader->load(dirname(__DIR__) . '/_data/robo2.yml'));
-        \Robo\Robo::config()->import($processor->export());
-
-        $argv = ['placeholder', 'test:simple-list'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        $this->guy->seeInOutput("a: '112'");
-        $this->guy->seeInOutput("b: '13'");
-    }
-
-    public function testCommandWithTaskConfiguration()
-    {
-        $loader = new YamlConfigLoader();
-        $loader->load(dirname(__DIR__) . '/_data/robo.yml');
-
-        \Robo\Robo::config()->import($loader->export());
-
-        $argv = ['placeholder', 'test:exec', '--simulate'];
-        $result = $this->runner->execute($argv, null, null, $this->guy->capturedOutputStream());
-
-        // `task.Exec.settings.dir` is defined in loaded robo.yml configuration file.
-        $this->guy->seeInOutput("->dir('/some/dir')");
     }
 }
